+++ src/main/java/seedu/classify/logic/commands/EditCommand.java
@@ -25,7 +25,7 @@ import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
 import seedu.classify.model.student.Student;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 /**
  * Edits the details of an existing student in the student record.
