+++ src/main/java/seedu/classify/logic/parser/AddStudCommandParser.java
@@ -20,7 +20,7 @@ import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
 import seedu.classify.model.student.Student;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 /**
  * Parses input arguments and creates a new AddStudCommand object
