+++ src/main/java/seedu/classify/logic/parser/ParserUtil.java
@@ -14,7 +14,7 @@ import seedu.classify.model.student.Email;
 import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
