+++ src/main/java/seedu/classify/model/util/SampleDataUtil.java
@@ -12,7 +12,7 @@ import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
 import seedu.classify.model.student.Student;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 /**
  * Contains utility methods for populating {@code StudentRecord} with sample data.
