+++ src/main/java/seedu/classify/storage/JsonAdaptedStudent.java
@@ -16,7 +16,7 @@ import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
 import seedu.classify.model.student.Student;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 /**
  * Jackson-friendly version of {@link Student}.
