+++ src/test/java/seedu/classify/logic/parser/AddStudCommandParserTest.java
@@ -45,7 +45,7 @@ import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
 import seedu.classify.model.student.Student;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 import seedu.classify.testutil.StudentBuilder;
 
 public class AddStudCommandParserTest {
