+++ src/test/java/seedu/classify/logic/parser/EditCommandParserTest.java
@@ -44,7 +44,7 @@ import seedu.classify.model.student.Email;
 import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 import seedu.classify.testutil.EditStudentDescriptorBuilder;
 
 public class EditCommandParserTest {
