+++ src/test/java/seedu/classify/logic/parser/ParserUtilTest.java
@@ -19,7 +19,7 @@ import seedu.classify.model.student.Class;
 import seedu.classify.model.student.Email;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 public class ParserUtilTest {
     private static final String INVALID_NAME = "R@chel";
