+++ src/test/java/seedu/classify/model/tag/ExamTest.java
@@ -3,6 +3,7 @@ package seedu.classify.model.tag;
 import static seedu.classify.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
+import seedu.classify.model.exam.Exam;
 
 public class ExamTest {
