+++ src/test/java/seedu/classify/testutil/EditStudentDescriptorBuilder.java
@@ -11,7 +11,7 @@ import seedu.classify.model.student.Id;
 import seedu.classify.model.student.Name;
 import seedu.classify.model.student.Phone;
 import seedu.classify.model.student.Student;
-import seedu.classify.model.tag.Exam;
+import seedu.classify.model.exam.Exam;
 
 /**
  * A utility class to help with building EditStudentDescriptor objects.
